move root files to dmsg in pkg
