+++ dmsg.go
@@ -1,0 +1,5 @@
+/*
+This file is a workaround to avoid go module errors.
+*/
+
+package dmsg
