+++ examples/basics/main.go
@@ -7,8 +7,8 @@ import (
 
 	"golang.org/x/net/nettest"
 
-	"github.com/skycoin/dmsg"
 	"github.com/skycoin/dmsg/pkg/disc"
+	dmsg "github.com/skycoin/dmsg/pkg/dmsg"
 
 	"github.com/skycoin/skywire-utilities/pkg/cipher"
 )
