+++ pkg/dmsg/const.go
