+++ pkg/dmsg/entity_common.go
