+++ pkg/dmsg/errors.go
