+++ pkg/dmsg/session_common.go
