+++ pkg/dmsg/stream.go
