+++ pkg/dmsg/stream_test.go
