+++ pkg/dmsg/util.go
