+++ pkg/dmsghttp/http_transport_test.go
@@ -12,7 +12,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/nettest"
 
-	"github.com/skycoin/dmsg"
+	"github.com/skycoin/dmsg/pkg/dmsg"
+
 	"github.com/skycoin/dmsg/pkg/disc"
 
 	"github.com/skycoin/skywire-utilities/pkg/cipher"
