+++ pkg/dmsgpty/host_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/skycoin/skycoin/src/util/logging"
 
-	"github.com/skycoin/dmsg"
+	dmsg "github.com/skycoin/dmsg/pkg/dmsg"
 	"github.com/skycoin/dmsg/pkg/dmsgtest"
 
 	"github.com/skycoin/skywire-utilities/pkg/cipher"
