+++ pkg/dmsgtest/env.go
@@ -9,8 +9,8 @@ import (
 
 	"golang.org/x/net/nettest"
 
-	"github.com/skycoin/dmsg"
 	"github.com/skycoin/dmsg/pkg/disc"
+	dmsg "github.com/skycoin/dmsg/pkg/dmsg"
 
 	"github.com/skycoin/skywire-utilities/pkg/cipher"
 )
